% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Panjer.R
\name{Panjer.Poisson}
\alias{Panjer.Poisson}
\title{Calculation of cumulative loss by Panjer recursion.}
\usage{
Panjer.Poisson(p, lambda)
}
\arguments{
\item{p}{the probability distribution (NumericVector)}

\item{lambda}{The parameters of the poisson distribution (numeric)}
}
\value{
The maximum possible cumulative loss and the probability of each value (list)
}
\description{
The prediction model is described in <Modern actuarial risk theory - based on R>.
}
\examples{
\dontrun{
p <- c(0.25,0.5,0.25)
lambda <- 4
fs <- Panjer.Poisson(p,lambda)
f <- Panjer.Poisson(p,lambda)$f
s <- Panjer.Poisson(p,lambda)$s
}
}
