% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParaApp.R
\name{ParaApp}
\alias{ParaApp}
\title{Calculation of cumulative loss by parametric approximation.}
\usage{
ParaApp(s, mu, sigma, k, method)
}
\arguments{
\item{s}{the possible value of cumulative loss (numeric)}

\item{mu}{The mean of cumulative loss (numeric)}

\item{sigma}{The standard deviation of the cumulative loss (numeric)}

\item{k}{The skewness coefficient of cumulative loss (numeric)}

\item{method}{The method of parametric approximation (String)}
}
\value{
The probability that the cumulative loss is less than or equal to s (numeric)
}
\description{
The prediction model is described in <Risk model: insurance loss prediction based on R>.
}
\examples{
\dontrun{
s <- 1000
mu <- 300
sigma <- sqrt(60000)
k <- 1.2247
Fs1 <- ParaApp(s,mu,sigma,k,method="normal")
Fs2 <- ParaApp(s,mu,sigma,k,method="Tgamma")
Fs3 <- ParaApp(s,mu,sigma,k,method="normal power")
Fs4 <- ParaApp(s,mu,sigma,k,method="Wilson Hilfery")
}
}
