\name{SC19052-package}
\alias{SC19052-package}
\docType{package}
\title{
  This package provides several functions for collective risk model and risk measures
}
\description{
  There are three R functions which can compute the cumulative loss for collective risk model,
and a drawing function for risk measures which can draw the curve of VaR and TVaR changing 
with quantile level.
}
\author{
Hongfang Sun

Maintainer: Hongfang Sun <sunhf@mail.ustc.edu.cn>
}
\keyword{ package }
