% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaRplot.R
\name{VaRplot}
\alias{VaRplot}
\title{VaR and TVaR plot.}
\usage{
VaRplot(x)
}
\arguments{
\item{x}{The possible value of the random variable of loss (NumericVector)}
}
\value{
A plot
}
\description{
The prediction model is described in <Risk model: insurance loss prediction based on R>.
}
\examples{
\dontrun{
x <- rnorm(100,2,1000)
VaRplot(x)
}
}
